Move to internal database drivers
Move from 3rd party db drivers to drivers supporting transactions in newer versions of Django. We are targeting Django 1.11.29 in this initial step. This removes the dependency mentioned in the project level README:

- django-transaction-hooks <http://django-transaction-hooks.readthedocs.org/>
  (needed only until EZID has been migrated to Django 1.9);

Move to internal database drivers

Move from 3rd party db drivers to drivers supporting transactions in newer versions of Django. We are targeting Django 1.11.29 in this initial step. This removes the dependency mentioned in the project level README:

- django-transaction-hooks <http://django-transaction-hooks.readthedocs.org/>
  (needed only until EZID has been migrated to Django 1.9);