Refactor app startup config methodology
EZID was using a system where modules configured themselves as a side effect of being imported. The method caused issues with wrong order of configuration. This refactor gives us control over when and how modules are configured.

Configuration of the app has been delegated to a new method that is called by Django after all apps have been loaded, and which configures the modules in an order independent of order of imports.